backend/api/requirements.txt
aiokafka==0.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.124.4
greenlet==3.3.0
h11==0.16.0
idna==3.11
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
